MindMorphix0005/brain-backend-Model/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
